Analytics chart removed from API Platform dashboard (landing page)
As part of the removal of Analytics chart, also the way to go to Analytics site has changed. Now the user has to select *Analytics* on the top right menu. So the image under `Accessing the Analytics Dashboards` has to be updated.
I send an image to Nicolas Earnshaw that can be used to replace the mentioned one, if team feels that meets the docs standards.